requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth
google-auth-httplib2
pydantic
